Code for generating influencing factors
The code uses multiple libraries to compute influencing factor raster data. Because of the large size of the raster data, Dask was used to compute in chunks. This operation is not necessary either if the RAM size is sufficient, or the raster files are small enough to fit into Memory